backend/requirements.txt
fastapi==0.99.1
uvicorn==0.22.0
transformers==4.30.2
torch==2.0.1
accelerate==0.20.3
pytest==7.4.0
pandas==2.0.3
scikit-learn==1.3.0
sentence-transformers==2.2.2
peft==0.3.0
rank-bm25==0.2.2
datasets==2.13.1
apache-airflow==2.2.3
MarkupSafe==2.0.1
python-dotenv==1.0.0